Focus cursor on text box on load without showing dropdown (#205)
Set focus on the search text box on load, but without showing the
dropdown the first time. Don't focus on mobile because the keyboard
takes up half the screen.

The text box autofocus was added in 3be7810 but removed in 4967561
because the focus caused the dropdown to show which blocked other text.



https://github.com/user-attachments/assets/0b70004c-65d8-4384-9935-460045092594



https://github.com/user-attachments/assets/c1bbd2a6-167e-4b61-a2c6-169b28d0dd2b